requirements.txt
appdirs==1.4.4
asgiref @ file:///tmp/build/80754af9/asgiref_1617740417847/work
black==21.6b0
certifi==2020.12.5
charset-normalizer==2.0.3
click==8.0.3
Django @ file:///tmp/build/80754af9/django_1620309509374/work
django-cors-headers==3.7.0
django-filter==21.1
djangorestframework==3.12.4
idna==3.2
Markdown==3.3.4
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2 @ file:///tmp/build/80754af9/psycopg2_1612298595717/work
pycosat==0.6.3
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
regex==2021.4.4
requests==2.26.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
toml==0.10.2
urllib3==1.26.6
